Moves delivery to internal controller
